Split libraries
